Fix several clang issues

+++ src/util/ScaLapackMatrix.cxx
@@ -58,7 +58,10 @@ template <typename F>
 ScaLapackMatrix<F>::ScaLapackMatrix(
   CTF::Matrix<F> &A, BlacsWorld *blacsWorld_, int blockSize
 ):
-  ScaLapackDescriptor(blacsWorld_, std::array<int,2>{{A.lens[0],A.lens[1]}}.data(), blockSize),
+  ScaLapackDescriptor(blacsWorld_,
+                      std::array<int,2>{{static_cast<int>(A.lens[0]),
+                                           static_cast<int>(A.lens[1])}}.data(),
+                      blockSize),
   blacsWorld(blacsWorld_)
 {
   // allocate local data
